pyproject.toml
[project]
name = "mcp-mastery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.1.0",
    "ipython==9.10.0",
    "langchain-groq==1.1.2",
    "mcp-use==1.6.0",
    "pypdf==6.7.5",
    "python-dotenv==1.2.2",
    "requests==2.32.5",
]
